chore: moving test around, renaming them
