+++ app-backend/libs/embeds/embeds-process.test.js
@@ -1,4 +1,4 @@
-import { parse, process } from "../";
+import { parse, process } from ".";
 
 describe("Embed processor", () => {
   it("should return empty array in case nothing is detected", () => {
