+++ app-backend/libs/embeds/video/basic.test.js
@@ -1,4 +1,4 @@
-import Video from "../basic";
+import Video from "./basic";
 
 const extensions = ["ogv", "webm", "mp4", "mov"];
 let embed;
